Move the Git clone flow into GitCoordinator
The git.clone UI moves into GitCoordinator, alongside the engine it
already used (GitService.clone + the gitError dialog). The first slice of
the remaining "Git UI wiring" follow-up to the Git-operations extraction.

Moved: the URL+destination OverlayInput form (the <home>/<repo> auto-fill +
the Browse picker), the git clone call, openClonedEntry (open a file from
the fresh clone so Git lights up), and the pure/unit-tested repoNameFromUrl
helper. The form method is cloneRepo() — renamed off clone() because a
no-arg clone() clashes with Object.clone()'s return type.

GitCoordinator.WindowOps gains keymap() (for the form's TextInputKeymap
caret nav) + openPath() (open the clone's README). MainController's three
clone callbacks — the git.clone command, the empty-GitPanel button, and
the BranchPopup No-VCS action — delegate to git::cloneRepo.
MainControllerTest.repoNameFromUrl -> GitCoordinator.repoNameFromUrl.

MainController: 8847 -> 8716 lines (-131). Behavior-preserving (the form/
clone/open bodies moved verbatim with mechanical host/ops substitutions).

Verified: full `mvn verify` green (spotless + JaCoCo + 1178 tests). Adds a
GitStateFxTest case that drives cloneRepo() on a real window and asserts
the clone form renders as an in-scene overlay through the coordinator's
CoordinatorHost/WindowOps wiring (no network).
