+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Internal: moved the Git clone flow into `GitCoordinator`.** The `git.clone` UI — the URL + destination `OverlayInput` form (with the `<home>/<repo>` auto-fill and the Browse picker), the `git clone` call, and opening a file from the fresh clone (`openClonedEntry`) — moved into `GitCoordinator` alongside the engine it already used (`GitService.clone`/`gitError`), as `cloneRepo()` (renamed off `clone` to avoid clashing with `Object.clone()`). The pure/unit-tested `repoNameFromUrl` helper moved with it (`GitCoordinator.repoNameFromUrl`). `MainController`'s three clone callbacks — the `git.clone` command, the empty-`GitPanel` button, and the `BranchPopup` No-VCS action — now delegate to `git::cloneRepo`. `MainController` shed ~131 lines (8,847 → ~8,716). Verified by `GitStateFxTest` (a new case asserts the clone form renders as an in-scene overlay through the coordinator wiring) + the moved unit tests. No behavior change — part of the ongoing `MainController` decomposition (the remaining Git-Log / blame / branch-popup wiring is a follow-up).
 - **Internal: extracted `RemoteCoordinator` from `MainController`.** The Remote file access (SFTP) feature — the SSH connect form, the saved-connection picker, mounting a remote folder as the Project tree root, single-file open from an `sftp://` URI, and disconnect — moved into its own `CoordinatorHost`-style coordinator. It owns the `RemoteFileSystems` engine (the SSH client + `Vfs` resolver), the `RemoteConnectionsPanel`, and the mount state; `MainController` keeps the `remote` `ToolWindow` (built with the coordinator's panel), queries the coordinator for the current-folder view, and wires the `remote.*`/`tool.remote` commands + the Welcome-page quick-connect callback to it. `MainController` shed ~172 lines (9,019 → ~8,847). Verified by a new FX test that drives the connect flow on a real window and asserts the SFTP form renders as an in-scene overlay (no network). No behavior change — part of the ongoing `MainController` decomposition.
 - **Internal: extracted `PluginCoordinator` from `MainController`.** The Plugin support feature — discovery/apply/install/uninstall plus the per-window `PluginContext` adapter that lets a plugin contribute commands, keybindings, tool windows, status-bar segments, and editor menu items — moved into its own `CoordinatorHost`-style coordinator. It owns the per-window `PluginRegistry`/`PluginInstaller`, the started-plugin instances, the plugin-contributed tool windows + editor menu items, the "Browse Plugins" picker, and the two inner adapters (`PluginContextImpl`/`ActiveEditorImpl`). The shared `PluginManager` (owned by `WindowManager`) is passed in; `MainController` keeps the `pluginManager` field + thin `disposePlugins()`/`pluginsEnabled()` delegates and wires `applyPlugins` / the tool-window gating / the editor-menu contributor / the `plugins.*`+`view.togglePlugins` commands to the coordinator. `MainController` shed ~574 lines (9,578 → ~9,004). Device-tested live (MCP bridge): the example plugin's Java-registered command, its tool-window toggle, and its declarative external command all loaded + ran through the moved `PluginContextImpl.start` path. No behavior change — part of the ongoing `MainController` decomposition (the largest single extraction in the campaign).
 - **Internal: extracted `HistoryCoordinator` from `MainController`.** The Local File History feature (per-file save/label/delete snapshots stored outside the file) moved into its own `CoordinatorHost`-style coordinator: it owns the `HistoryService` (off-thread snapshot/blob engine) + the `FileHistoryPanel`, and the whole flow — record (save / autosave / external / pre-delete), the per-file + folder-history views, label / recent-changes pickers, and revision diff/restore (reusing the shared `DiffCoordinator`). `MainController` keeps the File History `ToolWindow` (built with the coordinator's panel) and wires the record hooks + the `history.*`/`tool.fileHistory` commands + the project-tree "Show Local History" item to it. `MainController` shed ~300 lines (9,881 → ~9,578). Device-tested live (MCP bridge): saving a file recorded a revision into `history/index.json`, and the history commands ran clean. No behavior change — part of the ongoing `MainController` decomposition.
